refactor: extract crc32 and rle into community modules
Move base/crc32.{c,h} -> modules/polykybd/crc32/ and base/rle.{c,h} ->
modules/polykybd/rle/, each with a qmk_module.json so they ride QMK's
community-modules build path. Register both in corne42/keyboard.json and
split72/keyboard.json, drop the explicit SRC entries from rules.mk, and
update the half-dozen include sites to drop the now-redundant "base/"
prefix.

The crc32 lookup table is now static const (previously a writable
global) — moves ~1 KB from RAM to flash and keeps the symbol contained
inside the module, which is the whole point of encapsulating it.

rle.h gains documentation of the on-the-wire encoding and the
volatile-qualified compressed pointer rationale; behaviour of
rle_decompress / rle_count is unchanged. Surveyed the call sites
(fill_overlay.c, overlay.c::copy_rectangle_to_overlay, multicore_exec.c)
for additional RLE utilities to bundle but the surrounding helpers are
overlay-/ROI-/SCREEN_WIDTH-specific, not pure RLE — left in place.
