+++ modules/polykybd/crc32/crc32.c
@@ -3,7 +3,7 @@
 
 #include "crc32.h"
 
-uint32_t crc32Lookup[256] =
+static const uint32_t crc32Lookup[256] =
 { 0x00000000,0x77073096,0xEE0E612C,0x990951BA,0x076DC419,0x706AF48F,0xE963A535,0x9E6495A3,
   0x0EDB8832,0x79DCB8A4,0xE0D5E91E,0x97D2D988,0x09B64C2B,0x7EB17CBD,0xE7B82D07,0x90BF1D91,
   0x1DB71064,0x6AB020F2,0xF3B97148,0x84BE41DE,0x1ADAD47D,0x6DDDE4EB,0xF4D4B551,0x83D385C7,
@@ -52,4 +52,3 @@ uint32_t crc32_1byte(const void* data, uint16_t length, uint32_t previousCrc32)
 
   return ~crc; // same as crc ^ 0xFFFFFFFF
 }
-
