+++ modules/polykybd/crc32/crc32.h
