+++ modules/polykybd/crc32/qmk_module.json
@@ -1,0 +1,6 @@
+{
+    "module_name": "CRC32",
+    "maintainer": "PolyKybd",
+    "license": "GPL-2.0-or-later",
+    "url": "https://github.com/thpoll83/qmk_firmware"
+}
