+++ modules/polykybd/rle/qmk_module.json
@@ -1,0 +1,6 @@
+{
+    "module_name": "RLE",
+    "maintainer": "PolyKybd",
+    "license": "GPL-2.0-or-later",
+    "url": "https://github.com/thpoll83/qmk_firmware"
+}
