+++ modules/polykybd/rle/rle.c
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
+
 #include "rle.h"
 
 #include <stdbool.h>
